+++ tomcat/src/main/java/jakarta/Cookie.java
@@ -1,4 +1,4 @@
-package org.apache.catalina;
+package jakarta;
 
 import org.apache.coyote.http11.HttpCookie;
 import org.apache.coyote.http11.constants.HttpCookies;
