+++ tomcat/src/main/java/jakarta/ViewModel.java
@@ -1,5 +1,7 @@
-package org.apache.catalina;
+package jakarta;
 
+
+import org.apache.catalina.Session;
 
 public record ViewModel(String path, Object model, Cookie cookie, Session session) {
